+++ NEWS.md
@@ -13,6 +13,7 @@
 * `perc()`: Modified.  Added ability to use "and equals" or not to the items in `dir=` (i.e., there are now four items in `dir=`).  Fixed a bug related to using `na.rm=FALSE` and a "less than" situation.  Added some checks with error messages.  Added suite of tests.
 * `pcumsum()`: Modified.  Added some checks with error messages.  Added suite of tests.
 * `rcumsum()`: Modified.  Added some checks with error messages.  Added suite of tests.
+* `se()`: Modified.  Added some checks with error messages.  Added suite of tests.
 
 # FSA 0.6.13 May15
 * Some miscellaneous reorganizations of files.